refactor: move printer modules to reporter package
