+++ src/builtin-reporters/summary/diagnostic-items.test.ts
@@ -2,7 +2,7 @@ import assert from "node:assert";
 import { describe, it } from "node:test";
 import { mockWritable } from "../../../test/test-utils.ts";
 import type { RichDiagnostics } from "../../diagnostics/index.ts";
-import { createPrinter } from "../../printer/printer.ts";
+import { createPrinter } from "../../reporter/index.ts";
 import { collectRichDiagnosticItems, printRichDiagnosticItems } from "./diagnostic-items.ts";
 
 describe("builtin-reporters/summary/diagnostic-items", () => {
