+++ src/index.ts
@@ -115,7 +115,12 @@ export type {
 	NodeInspector,
 	ResultsBuilder,
 } from "./inspector/index.ts";
-export { createPrinter, type Printer, type PrinterOptions } from "./printer/printer.ts";
-export { createXmlPrinter, type XmlPrinter } from "./printer/xml-printer.ts";
 export type { Reporter } from "./reporter/index.ts";
+export {
+	createPrinter,
+	createXmlPrinter,
+	type Printer,
+	type PrinterOptions,
+	type XmlPrinter,
+} from "./reporter/index.ts";
 export type { TsInspectError };
