+++ src/reporter/index.ts
@@ -2,4 +2,6 @@
  * Reporter module exports.
  */
 
+export { createPrinter, type Printer, type PrinterOptions } from "./printer/printer.ts";
+export { createXmlPrinter, type XmlPrinter } from "./printer/xml-printer.ts";
 export type { Reporter } from "./reporter.ts";
